Fix graphviz detection in Network_conti and inline NetGraph helpers
Network_conti.py called importlib.util.find_spec and importlib.util.module_from_spec
but imported only the top-level importlib package. importlib.util is a submodule
that is not bound on the importlib object unless explicitly imported, so this
relied on a side-effect import happening elsewhere first and raised
AttributeError on import. Detect graphviz with a plain try/except import
instead, which is the standard idiom for an optional dependency and avoids the
importlib.util pitfall entirely.

NetGraph.py and NetGraph_empty.py each defined a single function, Drawnet_notag,
selected at import time by whether graphviz was installed: the real
implementation or a stub that always raised ModuleNotFoundError. Inline both into
Network_conti.py as one Drawnet_notag that raises ModuleNotFoundError up front
when graphviz could not be imported, removing the two single-purpose files and
the spec/module_from_spec indirection used to pick between them.
