--- cytnx/NetGraph_empty.py
@@ -1,4 +1,0 @@
-
-def Drawnet_notag(opt_name,tn_names, edges_info,dangling_infos,figsize=[6,5],engine=''):
-    raise ModuleNotFoundError("[ERROR] graphviz is not installed!")
-    pass
